Fixed ouput number issue

+++ tests/data_gen.py
@@ -18,6 +18,8 @@ def dir_gen(directory_name):
     """
     if not os.path.exists(directory_name):
         os.makedirs(directory_name)
+        return True
+    return False
 
 
 def file_gen(file_name):
@@ -26,9 +28,12 @@ def file_gen(file_name):
 
     :param string filename: Fully qualified file name which is to be craeted
     """
-    file_hanlder = open(file_name, 'wb+')
-    file_hanlder.write(b'This is a test file')
-    file_hanlder.close()
+    if not os.path.exists(file_name):
+        file_hanlder = open(file_name, 'wb+')
+        file_hanlder.write(b'This is a test file')
+        file_hanlder.close()
+        return True
+    return False
 
 
 def millis_time():
@@ -55,8 +60,9 @@ def main():
         if random.choice(is_directory):
             choice = random.choice(name_list)
             dir_name = os.path.join(*dir_list, choice)
-            dir_gen(dir_name)
-            dir_count += 1
+            check = dir_gen(dir_name)
+            if check:
+                dir_count += 1
             if random.choice(go_inside) and len(dir_name) < max_path_length:
                 dir_list.append(choice)
             elif (
@@ -70,8 +76,9 @@ def main():
             name_part = random.choice(name_list)
             full_name = name_part + '_' + millis_time() + '.' + extn
             file_name = os.path.join(*dir_list, full_name)
-            file_gen(file_name)
-            file_count += 1
+            check = file_gen(file_name)
+            if check:
+                file_count += 1
 
     print('Created {} files and {} directories in {}'
           .format(file_count, dir_count, parent_directory))
